pyproject.toml
# Licensed to the Apache Software Foundation (ASF) under one
# or more contributor license agreements.  See the NOTICE file
# distributed with this work for additional information
# regarding copyright ownership.  The ASF licenses this file
# to you under the Apache License, Version 2.0 (the
# "License"); you may not use this file except in compliance
# with the License.  You may obtain a copy of the License at
#
#   http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing,
# software distributed under the License is distributed on an
# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
# KIND, either express or implied.  See the License for the
# specific language governing permissions and limitations
# under the License.

[project]
name = "openserverless-admin-api"
version = "0.1.0"
description = "OpenServerless Admin Api"
authors = [
    { name = "Michele Sciabarra", email = "michele@sciabarra.com" },
    { name = "Francesco Timperi Tiberi", email = "francesco@nuvolaris.io" },
    { name = "Bruno Salzano", email = "d4kstar@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "PyYAML>=6.0.2",
    "Jinja2>=3.1.3",
    "requests>=2.31.0",
    "flatdict==4.0.1", # è l’ultima disponibile su PyPI
    "minio==7.1.13",
    "backoff==2.2.1", # ultima versione disponibile
    "psycopg-binary==3.1.18",
    "psycopg==3.1.18",
    "flask>=3.1.1",
    "redis>=5.2.0",
    "auth0-python>=4.9.0",
    "waitress==3.0.2",
    "flasgger==0.9.7.1",
    "flask-cors>=6.0.0",
    "cryptography>=44.0.0",
    "python-dotenv>=1.1.1",
    "bcrypt>=4.3.0",
]

[project.optional-dependencies]
dev = [
        "ipython>=8.24.0"
    ]

[project.scripts]
system = "nuvolaris.app:main"
